Extract shared text extraction utilities to core module
Move classify_mimetype, extract_text, extract_pdf_text, extract_docx_text,
and related helpers from integrations/odoo/tools/pdf_tools.py to
core/agents/tools/text_extraction.py. Both Gmail attachment tools and
Odoo attachment tools now import from the shared location, eliminating
the cross-integration coupling where Gmail depended on Odoo imports.
